+++ ModularFeedTDDCaseStudy/ModularFeedTDDCaseStudy/FeedCache/LocalFeedLoader.swift
@@ -8,25 +8,6 @@
 
 import Foundation
 
-// never can have an instance of a 'FeedCachePolicy'
-// because it needs no identity and it holds no state
-private final class FeedCachePolicy {
-    private init() {}
-    
-    private static let calendar = Calendar(identifier: .gregorian)
-    
-    private static var maxCacheAgeInDays: Int {
-        return 7
-    }
-    
-    static func validate(_ timestamp: Date, against date: Date) -> Bool {
-        guard let maxCacheAge = calendar.date(byAdding: .day, value: maxCacheAgeInDays, to: timestamp) else {
-            return false
-        }
-        return date < maxCacheAge
-    }
-}
-
 public final class LocalFeedLoader {
     private let store: FeedStore
     private let currentDate: () -> Date
